chore(next.config.js): add sassOptions to nextConfig to include the path to the src/styles directory chore(package.json): add sass as a devDependency feat(components): add primary-button component feat(components/primary-button): create PrimaryButton component with styles feat(components/primary-button/primary-button.module.scss): add styles for PrimaryButton component chore(pages/_app.tsx): update import statement for global styles from .css to .scss chore(pages/index.tsx): replace existing content with PrimaryButton component chore(styles/Home.module.css): delete unused file The changes were made to improve the project's styling and structure. The addition of the sassOptions in next.config.js allows the project to use Sass and include the path to the src/styles directory. The new PrimaryButton component and its associated styles provide a reusable button component with a specific style. The import statement in pages/_app.tsx was updated to reflect the change from .css
feat(styles): add _reset.scss file to reset default styles and improve consistency
feat(styles): add _variables.scss file to define common color and font variables
feat(styles): add sf-pro.ttf font file for custom font usage
refactor(styles): remove globals.css file as it is no longer needed
The _reset.scss file is added to reset default styles and improve consistency across different browsers. The _variables.scss file is added to define common color and font variables for easier theming. The sf-pro.ttf font file is added to provide a custom font option. The globals.css file is removed as it is no longer needed after the introduction of the _reset.scss file.

feat(styles): add global styles in globals.scss file
chore(tsconfig.json): add paths for @components and @components/* in tsconfig.json
The globals.scss file is added to the project, which contains global styles for the application. This file imports the "reset" and "variables" stylesheets and defines the font-face for the "SF Pro" font. It also sets the font-family, color, and font-size for the html and body elements, and sets the background-color for the #__next element. Additionally, it sets the overflow-x property to hidden for the body element.

In the tsconfig.json file, paths are added for the @components and @components/* aliases. This allows for easier import of components in the project by using these aliases instead of specifying the full path.